refactor: 1단계 피드백 반영
- 요구사항 문맥에 맞게 변경
  - 제약사항 : ~한다
  - 가능한 행위 : ~할 수 있다
- 테스트용 리소스 생성 역할의 네이밍을 `TestFixture` 로 변경
- 테스트 코드의 setup 메서드 내에서 조건 설정하는 것을 분리
